+++ CHANGELOG.md
@@ -1,3 +1,71 @@
+## [0.21.0] - 2026-07-06
+
+### 🚀 Features
+
+- *(types)* Add rwaMarketType to Market ([#111](https://github.com/backpack-exchange/bpx-api-client/issues/111))
+## [0.20.6] - 2026-07-01
+
+### 🐛 Bug Fixes
+
+- *(types)* Skip None fields in FillsHistoryParams query serialization ([#109](https://github.com/backpack-exchange/bpx-api-client/issues/109))
+## [0.20.5] - 2026-06-30
+
+### 🚀 Features
+
+- *(types)* Expose market order book state and add unknown enum fallbacks ([#108](https://github.com/backpack-exchange/bpx-api-client/issues/108))
+## [0.20.4] - 2026-06-18
+
+### 🚀 Features
+
+- Add dependabot ([#91](https://github.com/backpack-exchange/bpx-api-client/issues/91))
+
+### ⚙️ Miscellaneous Tasks
+
+- *(types)* Remove serde and strum rename attributes for blockchain ([#107](https://github.com/backpack-exchange/bpx-api-client/issues/107))
+- Bump reqwest, tokio-tungstenite, serde_qs and strum versions ([#104](https://github.com/backpack-exchange/bpx-api-client/issues/104))
+## [0.20.3] - 2026-06-07
+
+### 🐛 Bug Fixes
+
+- Updates to the securities endpoint ([#106](https://github.com/backpack-exchange/bpx-api-client/issues/106))
+## [0.20.1] - 2026-04-21
+
+### 🚀 Features
+
+- Add remaining_mint_quantity to Vault struct
+- Added get_vault_mints and get_vault_redeems
+- Add the option to set a timeout in the BpxClientBuilder ([#92](https://github.com/backpack-exchange/bpx-api-client/issues/92))
+- Add security endpoints ([#105](https://github.com/backpack-exchange/bpx-api-client/issues/105))
+
+### 🚜 Refactor
+
+- Remove get_vault_pending_redeems which is admin-only
+## [0.18.0] - 2026-03-02
+
+### 🚀 Features
+
+- Add support for current vault endpoints
+- Add public method base_url to BpxClient
+- Add public method execute to BpxClient
+
+### 🚜 Refactor
+
+- Expose public method build_signed_request
+## [0.17.4] - 2026-02-10
+
+### 🚀 Features
+
+- Vault pending redeems query ([#100](https://github.com/backpack-exchange/bpx-api-client/issues/100))
+## [0.17.1] - 2026-01-23
+
+### 🚀 Features
+
+- Parse system order type on RFQs ([#97](https://github.com/backpack-exchange/bpx-api-client/issues/97))
+## [0.16.0] - 2026-01-21
+
+### 🐛 Bug Fixes
+
+- Fix incorrect type on trigger_quantity field on OrderUpdate struct ([#93](https://github.com/backpack-exchange/bpx-api-client/issues/93))
 ## [0.15.1] - 2025-12-15
 
 ### 🐛 Bug Fixes